// src/app/views/interfaces/profileUser.ts
export class ProfileUser {
    administrator?: boolean;

    createdAt?: string;

    functions?: string;

    governorateMunicipality?: string;

    img?: string;

    is_active?: boolean;

    lastname?: string;

    name?: string;

    phone?: string;

    position?: string;

    profession?: string;

    rol?: string;

    updatedAt?: string;
}